Move more of the search code out of GameBrowser
